<docs)>(blog): Update swag document
Updates the swagger document for the api endpoint.
The two approaches to describe multiple request body parameters in swagger documentation are either to write them directly in the comments or define a separate structure to avoid errors of multiple bodies. Although it is possible to use multiple bodies, it may lead to errors in the `.yaml ` file in editor.swagger.io.

#3