<feat>: Add middleware for downstream API calls
This commit adds a middleware named ContextTimeout for setting timeout controls when making downstream API calls.  i also write corresponding test cases to ensure the middleware works properly.

In addition ,we add a codecov.io badge to show the coverage of the project ,which helps other developers understand the testing situation of the project .
refs #15,#8