<feat>: Add functionality to start with follow-up parameters using the flag package
Enhance the NewSetting function to utilize the flag package for parsing command-line arguments. This update allows the function to accept a "-config" flag representing the path to the configuration files. If no "-config" flag is provided, it defaults to using the "configs/" path. This improvement provides a more flexible and customizable way to specify configuration paths when starting the application.
#10